backend/requirements.txt
# Core Django and API Dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# ML and Computer Vision Dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.0
scipy>=1.11.0

# Data Processing and Validation
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance Monitoring and Logging
psutil>=5.9.0
GPUtil>=1.4.0
prometheus-client>=0.17.0

# Image Processing and Utilities
imageio>=2.31.0
opencv-contrib-python>=4.8.0
Shapely>=2.0.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0
black>=23.0.0
flake8>=6.0.0

# Production Deployment
gunicorn>=21.2.0
whitenoise>=6.5.0
celery>=5.3.0
redis>=4.6.0

# Optional: CUDA support (uncomment if using GPU)
# torch-audio>=2.0.0+cu118
# torchaudio>=2.0.0+cu118

# Database (SQLite included with Django, PostgreSQL for production)
# psycopg2-binary>=2.9.0  # Uncomment for PostgreSQL